% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceptanceRate}
\alias{acceptanceRate}
\title{Computing acceptance rate based on trace
Note: Only use for traces from continous distributions!}
\usage{
acceptanceRate(trace)
}
\description{
Computing acceptance rate based on trace
Note: Only use for traces from continous distributions!
}
\keyword{internal}
