% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acvBlockMatrix}
\alias{acvBlockMatrix}
\title{Build an nd times nd Block Toeplitz matrix from the
(d times d) autocovariances gamma(0),...,gamma(n-1)}
\usage{
acvBlockMatrix(acv)
}
\description{
Build an nd times nd Block Toeplitz matrix from the
(d times d) autocovariances gamma(0),...,gamma(n-1)
}
\keyword{internal}
