% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acvMatrix}
\alias{acvMatrix}
\title{Build an n times n Toeplitz matrix from the 
autocovariance values gamma(0),...,gamma(n-1)}
\usage{
acvMatrix(acv)
}
\description{
Build an n times n Toeplitz matrix from the 
autocovariance values gamma(0),...,gamma(n-1)
}
\keyword{internal}
