% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_pacf.R
\name{ar_lik}
\alias{ar_lik}
\title{Full likelihood of an autoregressive time series model with i.i.d. normal innovations}
\usage{
ar_lik(x, ar, mean = 0, sd = 1, log = F)
}
\arguments{
\item{x}{numeric vector of data}

\item{ar}{vector of ar parameters}

\item{mean}{the innovation mean}

\item{sd}{the innovation standard deviation}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
numeric value for the likelihood or log-likelihood
}
\description{
Full likelihood of an autoregressive time series model with i.i.d. normal innovations
}
