% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{arma_conditional}
\alias{arma_conditional}
\title{Negative ARMA(p, q) log likelihood}
\usage{
arma_conditional(zt, ar, ma, nll_fun, full_lik, ...)
}
\description{
Negative ARMA(p, q) log likelihood
}
\keyword{internal}
