% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_S3method.R
\name{bayes_factor.bdp_dw_result}
\alias{bayes_factor.bdp_dw_result}
\title{Extracting the Bayes factor of k1=1 from bdp_dw_result class}
\usage{
\method{bayes_factor}{bdp_dw_result}(obj)
}
\arguments{
\item{obj}{object of class bdp_dw_result}
}
\value{
the estimated Bayes factor of k1=1
}
\description{
Extracting the Bayes factor of k1=1 from bdp_dw_result class
}
