% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_S3method.R
\name{bayes_factor}
\alias{bayes_factor}
\title{a generic method for bdp_dw_result class}
\usage{
bayes_factor(obj)
}
\arguments{
\item{obj}{object of class bdp_dw_result}
}
\description{
a generic method for bdp_dw_result class
}
