% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_prior_and_mcmc_params.R
\name{bdp_dw_bayes_factor_k1}
\alias{bdp_dw_bayes_factor_k1}
\title{Estimating the Bayes factor of hypothesis "k1 = 1".}
\usage{
bdp_dw_bayes_factor_k1(post_sample, precision = 1000)
}
\arguments{
\item{post_sample}{the posterior sample generated by \link{bdp_dw_mcmc}.}

\item{precision}{a positive integer specifying the number of terms used in approximating
the normalizing constant of the prior probability mass function of k1. Default 1000.}
}
\value{
The Savage-Dickey estimate of the Bayes factor and its theoretical upper bound. c.f. section 3.3 of Tang et al. (2023).
}
\description{
Estimating the Bayes factor of hypothesis "k1 = 1".
}
\references{
Tang et al. (2023)
\emph{Bayesian nonparametric spectral analysis of locally stationary processes}
ArXiv preprint
<arXiv:2303.11561>
}
