% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_prior_and_mcmc_params.R
\name{bdp_dw_mcmc_params_gen}
\alias{bdp_dw_mcmc_params_gen}
\title{Generate a list of values for MCMC algorithm}
\usage{
bdp_dw_mcmc_params_gen(
  Ntotal = 110000,
  burnin = 60000,
  thin = 10,
  adaptive.batchSize = 50,
  adaptive.targetAcceptanceRate = 0.44
)
}
\arguments{
\item{Ntotal}{total number of iterations to run the Markov chain}

\item{burnin}{number of initial iterations to be discarded}

\item{thin}{thinning number (for post-processing of the posterior sample)}

\item{adaptive.batchSize}{the batch size for the adaptive MCMC algorithm for sampling tau}

\item{adaptive.targetAcceptanceRate}{the target acceptance rate for the adaptive MCMC algorithm for sampling tau}
}
\value{
A list of MCMC parameter values
}
\description{
Generate a list of values for MCMC algorithm
}
