% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein_polynomials.R
\name{betaBasis_k}
\alias{betaBasis_k}
\title{Construct Bernstein polynomial basis of degree k on omega}
\usage{
betaBasis_k(omega, k, coarsened)
}
\arguments{
\item{omega}{numeric vector in [0,1] of evaluation points}

\item{k}{positive integer for the degree}

\item{coarsened}{bool flag indicating whether coarsened or standard Bernstein polynomials are used}
}
\description{
Construct Bernstein polynomial basis of degree k on omega
}
\keyword{internal}
