% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{Cn}
\alias{Cn}
\title{Function that calculates correction matrix C_n ^ (-1/2) for ARMA model}
\usage{
Cn(n, g)
}
\description{
Function that calculates correction matrix C_n ^ (-1/2) for ARMA model
}
\keyword{internal}

