% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein_polynomials.R
\name{coarsened_bernstein_i}
\alias{coarsened_bernstein_i}
\title{Helping function for \code{coarsened_bernstein}}
\usage{
coarsened_bernstein_i(omega, l, i)
}
\description{
Helping function for \code{coarsened_bernstein}
}
\keyword{internal}
