% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{densityMixture}
\alias{densityMixture}
\title{Construct a density mixture from mixture weights and density functions.}
\usage{
densityMixture(weights, densities)
}
\description{
Construct a density mixture from mixture weights and density functions.
}
\keyword{internal}
