% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{dtex.kurt}
\alias{dtex.kurt}
\title{Density of t-distribution in terms of excess kurtosis}
\usage{
dtex.kurt(x, ex.kurt)
}
\description{
Density of t-distribution in terms of excess kurtosis
}
\keyword{internal}

