% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_transform.R
\name{fast_ft}
\alias{fast_ft}
\title{Fast Fourier Transform}
\usage{
fast_ft(x, real = T)
}
\description{
Fast Fourier Transform
}
\details{
If \code{real}: computes F_n X_n with the real-valued Fourier 
transformation matrix F_n (see Section 2.1 in Kirch et al (2018)).
If \code{!real}: computes the complex-valued Fourier coefficients 
(see (4.5) in Meier (2018)).
}
\keyword{internal}
