% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_transform.R
\name{fast_ift}
\alias{fast_ift}
\title{Fast Inverse Fourier Transform}
\usage{
fast_ift(x, real = T, TOL = 1e-15)
}
\description{
Fast Inverse Fourier Transform
}
\details{
inverse function of \code{fast_ft}
}
\keyword{internal}
