% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_transform.R
\name{fourier_freq}
\alias{fourier_freq}
\title{Fourier frequencies}
\usage{
fourier_freq(n)
}
\arguments{
\item{n}{integer}
}
\value{
numeric vector of length floor(n/2)+1
}
\description{
Fourier frequencies on [0,pi], as defined by 2*pi*j/n for j=0,...,floor(n/2).
}
