% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genEpsARC}
\alias{genEpsARC}
\title{C++ function for generating epsilon process for AR(p)}
\usage{
genEpsARC(zt, ar)
}
\description{
C++ function for generating epsilon process for AR(p)
}
\keyword{internal}
