% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genEpsARMAC}
\alias{genEpsARMAC}
\title{Get epsilon process (i.e. model residuals) for ARMA(p,q)}
\usage{
genEpsARMAC(zt, ar, ma)
}
\description{
Get epsilon process (i.e. model residuals) for ARMA(p,q)
}
\keyword{internal}
