% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genEpsMAC}
\alias{genEpsMAC}
\title{C++ function for generating epsilon process for MA(q)}
\usage{
genEpsMAC(zt, ma)
}
\description{
C++ function for generating epsilon process for MA(q)
}
\keyword{internal}
