% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{generalizedGaussian.alpha}
\alias{generalizedGaussian.alpha}
\title{Help function for Generalized Gaussian}
\usage{
generalizedGaussian.alpha(beta)
}
\description{
Help function for Generalized Gaussian
}
\keyword{internal}
