% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_ar_algorithm.R
\name{gibbs_AR_nuisance_intern}
\alias{gibbs_AR_nuisance_intern}
\title{Gibbs sampler for Bayesian AR model in PACF parametrization,
including support for TS to be a nuisance parameter}
\usage{
gibbs_AR_nuisance_intern(data, mcmc_params, prior_params, model_params,
  full_lik = F)
}
\description{
Gibbs sampler for Bayesian AR model in PACF parametrization,
including support for TS to be a nuisance parameter
}
\keyword{internal}
