% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_vnp_algorithm.R
\name{gibbs_multivariate_nuisance}
\alias{gibbs_multivariate_nuisance}
\title{Gibbs sampler for corrected parametric likelihood + Bernstein-Dirichlet mixture,
including possibility of using time series as mere nuisance parameter
Multivariate case}
\usage{
gibbs_multivariate_nuisance(data, mcmc_params, corrected, prior_params,
  model_params)
}
\description{
Gibbs sampler for corrected parametric likelihood + Bernstein-Dirichlet mixture,
including possibility of using time series as mere nuisance parameter
Multivariate case
}
\keyword{internal}
