% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_np_algorithm.R
\name{gibbs_nuisance}
\alias{gibbs_nuisance}
\title{Gibbs sampler for corrected parametric likelihood + Bernstein-Dirichlet mixture,
including possibility of using time series as mere nuisance parameter}
\usage{
gibbs_nuisance(data, mcmc_params, corrected, prior_params, model_params,
  full_lik = NULL)
}
\description{
Gibbs sampler for corrected parametric likelihood + Bernstein-Dirichlet mixture,
including possibility of using time series as mere nuisance parameter
}
\keyword{internal}
