% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_pacf.R
\name{gibbs_pacf}
\alias{gibbs_pacf}
\title{Gibbs sampler of AR model with PACF parametrization}
\usage{
gibbs_pacf(data, Ntotal, burnin, Nadaptive = burnin,
  adaption.batchSize = 50, adaption.targetAcceptanceRate = 0.44, ar.order,
  sigma2.alpha = 0.001, sigma2.beta = 0.001, mu.prop = rep(0, ar.order),
  var.prop = rep(1/length(data), ar.order), psi.alpha = rep(1, ar.order),
  psi.beta = rep(1, ar.order))
}
\description{
Gibbs sampler of AR model with PACF parametrization
}
\keyword{internal}
