% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_var_algorithm.R
\name{gibbs_VAR_nuisance_intern}
\alias{gibbs_VAR_nuisance_intern}
\title{Gibbs sampling algorithm for VAR model}
\usage{
gibbs_VAR_nuisance_intern(data, mcmc_params, prior_params, model_params,
  full_lik = F)
}
\description{
Gibbs sampling algorithm for VAR model
}
\keyword{internal}
