% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hasEigenValueSmallerZero}
\alias{hasEigenValueSmallerZero}
\title{Does a matrix have an eigenvalue smaller than 0?}
\usage{
hasEigenValueSmallerZero(A, TOL = 0)
}
\description{
Does a matrix have an eigenvalue smaller than 0?
}
\keyword{internal}
