% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{l_generalizedGaussian}
\alias{l_generalizedGaussian}
\title{Help function for Generalized Gaussian}
\usage{
l_generalizedGaussian(x, beta)
}
\description{
Help function for Generalized Gaussian
}
\keyword{internal}

