% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_type.R
\name{lik_ar}
\alias{lik_ar}
\title{Likelihood of an autoregressive time series model with i.i.d. normal innovations}
\usage{
lik_ar(x, ar, mean = 0, sd = 1, log = F, full = T)
}
\arguments{
\item{x}{numeric vector of data}

\item{ar}{vector of ar parameters}

\item{mean}{the innovation mean}

\item{sd}{the innovation standard deviation}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}

\item{full}{logical; if TRUE, the full likelihood for all observations is computed; if FALSE, the partial likelihood for the last n-p observations}
}
\value{
numeric value for the likelihood or log-likelihood
}
\description{
Likelihood of an autoregressive time series model with i.i.d. normal innovations
}
\keyword{internal}
