% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_np_help.R
\name{llike}
\alias{llike}
\title{Log corrected parametric AR likelihood (Gaussian)}
\usage{
llike(omega, FZ, ar, ma, v, w, k, tau, corrected, toggle.q, pdgrm,
  beta_basis_k, nll_fun, f.alpha, excludeBoundary, full_lik)
}
\description{
Log corrected parametric AR likelihood (Gaussian)
}
\details{
See (5) in Kirch et al (2018)
}
\keyword{internal}
