% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_AR_nuisance.R
\name{llike_AR}
\alias{llike_AR}
\title{Time domain AR(p) likelihood for nuisance/noise time series}
\usage{
llike_AR(x_noise, rho, sigma2, full_lik)
}
\description{
Time domain AR(p) likelihood for nuisance/noise time series
}
\keyword{internal}
