% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_pacf.R
\name{llike_full}
\alias{llike_full}
\title{Full Gaussian likelihood (includes marginals; only needed for model comparison)}
\usage{
llike_full(data, psi, sigma2)
}
\description{
Full Gaussian likelihood (includes marginals; only needed for model comparison)
}
\keyword{internal}

