% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_pacf.R
\name{llike_pacf}
\alias{llike_pacf}
\title{Conditional Gaussian likelihood (good enough for sampling steps)}
\usage{
llike_pacf(data, psi, sigma2)
}
\description{
Conditional Gaussian likelihood (good enough for sampling steps)
}
\keyword{internal}
