% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{logDet_stickBreaking}
\alias{logDet_stickBreaking}
\title{Help function for proposing new values of v's}
\usage{
logDet_stickBreaking(v)
}
\description{
Help function for proposing new values of v's
}
\keyword{internal}
