% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logfuller}
\alias{logfuller}
\title{Fuller Logarithm}
\usage{
logfuller(x, xi = 1e-08)
}
\description{
Fuller Logarithm
}
\details{
see Fuller (1996), page 496
}
\references{
W. Fuller (1996)
\emph{Introduction to Statistical Time Series}
Wiley Series in Probability and Statistics
}
\keyword{internal}
