% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{lpost}
\alias{lpost}
\title{Log corrected posterior}
\usage{
lpost(omega, FZ, ar, ma, v, w, k, tau, M, g0.alpha, g0.beta, k.theta, tau.alpha,
  tau.beta, corrected, toggle.q, pdgrm, beta_basis_k, nll_fun, f.alpha, rho,
  rho.alpha, rho.beta, excludeBoundary, full_lik)
}
\description{
Log corrected posterior
}
\keyword{internal}
