% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_AR_nuisance.R
\name{lprior_AR}
\alias{lprior_AR}
\title{Log prior for PACF (beta) and sigma2 (Inverse Gamma), unnormalized}
\usage{
lprior_AR(rho, rho.alpha, rho.beta, sigma2, sigma2.alpha, sigma2.beta)
}
\description{
Log prior for PACF (beta) and sigma2 (Inverse Gamma), unnormalized
}
\keyword{internal}
