% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_pacf.R
\name{lprior_pacf}
\alias{lprior_pacf}
\title{Log-Prior of PACF parametrization}
\usage{
lprior_pacf(psi, sigma2, sigma2.alpha, sigma2.beta, psi.alpha, psi.beta)
}
\description{
Log-Prior of PACF parametrization
}
\keyword{internal}
