% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureWeight}
\alias{mixtureWeight}
\title{Get mixture weights of Bernstein-Dirchlet-Mixtures}
\usage{
mixtureWeight(p, w, k)
}
\description{
Get mixture weights of Bernstein-Dirchlet-Mixtures
}
\keyword{internal}
