% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_transform.R
\name{mpdgrm}
\alias{mpdgrm}
\title{Compute Periodgram matrix from (complex-valued) Fourier coefficients}
\usage{
mpdgrm(FZ)
}
\description{
Compute Periodgram matrix from (complex-valued) Fourier coefficients
}
\details{
see (4.7) in Meier (2018)
}
\keyword{internal}
