% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{my_ddirichlet_unnormalized}
\alias{my_ddirichlet_unnormalized}
\title{Help function for proposing new values of v's}
\usage{
my_ddirichlet_unnormalized(x, alpha, log)
}
\description{
Help function for proposing new values of v's
}
\keyword{internal}
