% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{my_rdirichlet}
\alias{my_rdirichlet}
\title{Help function for proposing new values of v's}
\usage{
my_rdirichlet(alpha)
}
\description{
Help function for proposing new values of v's
}
\keyword{internal}
