% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{nll_generalizedGaussian}
\alias{nll_generalizedGaussian}
\title{Help function for Generalized Gaussian}
\usage{
nll_generalizedGaussian(epsilon_t, beta)
}
\description{
Help function for Generalized Gaussian
}
\keyword{internal}
