% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nll_norm}
\alias{nll_norm}
\title{Negative log likelihood of iid standard normal observations [unit variance]
Note: deprecated}
\usage{
nll_norm(epsilon_t, ...)
}
\description{
Negative log likelihood of iid standard normal observations [unit variance]
Note: deprecated
}
\keyword{internal}
