% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{nll_norm_unnormalized}
\alias{nll_norm_unnormalized}
\title{unnormalized negative log likelihood of iid standard normal observations [unit variance]}
\usage{
nll_norm_unnormalized(epsilon_t, ...)
}
\description{
unnormalized negative log likelihood of iid standard normal observations [unit variance]
}
\keyword{internal}

