% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{nll_t}
\alias{nll_t}
\title{negative log likelihood of iid t observations with given excess kurtosis [unit variance]}
\usage{
nll_t(epsilon_t, ex.kurt, ...)
}
\description{
negative log likelihood of iid t observations with given excess kurtosis [unit variance]
}
\keyword{internal}

