% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{omegaFreq}
\alias{omegaFreq}
\title{Fourier frequencies, rescaled on the unit interval}
\usage{
omegaFreq(n)
}
\arguments{
\item{n}{integer length in consideration}
}
\value{
numeric vector of the rescaled Fourier frequencies
}
\description{
Fourier frequencies, rescaled on the unit interval
}

