% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pacf2AR}
\alias{pacf2AR}
\title{C++ function for computing AR coefficients, given PACF.}
\usage{
pacf2AR(pacf)
}
\description{
C++ function for computing AR coefficients, given PACF.
}
\keyword{internal}
