% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{pacfToAR}
\alias{pacfToAR}
\title{Convert partial autocorrelation coefficients to AR coefficients.}
\usage{
pacfToAR(pacf)
}
\arguments{
\item{pacf}{numeric vector of partial autocorrelations in (-1,1)}
}
\value{
numeric vector of autoregressive model coefficients
}
\description{
Convert partial autocorrelation coefficients to AR coefficients.
}
