% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_S3method.R
\name{plot.bdp_dw_result}
\alias{plot.bdp_dw_result}
\title{Plot method for bdp_dw_result class}
\usage{
\method{plot}{bdp_dw_result}(
  x,
  which = 1:4,
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  col = hcl.colors(200, "Blue-Red 3"),
  ...
)
}
\arguments{
\item{x}{object of class bdp_dw_result}

\item{which}{if a subset of the plots is required, specify a subset of the numbers 1:6. 1 indicates posterior mean,
2 indicates posterior median, 3 for lower bound of pointwise 90 percent credible region, 4 for upper bound of 
pointewise 90 percent credible region, 5 indicates lower bound of uniform 90 percent credible region, 6 indicates
upper bound of uniform 90 percent credible region.}

\item{ask}{logical; if TRUE, the user is asked before each plot.}

\item{col}{choice of color, default hcl.colors(200, "Blue-Red 3").}

\item{...}{other parameters to be passed through to \code{image.default}}
}
\description{
Plot method for bdp_dw_result class
}
