% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_S3method.R
\name{plot.bdp_dw_tv_psd}
\alias{plot.bdp_dw_tv_psd}
\title{Plot method for bdp_dw_tv_psd class}
\usage{
\method{plot}{bdp_dw_tv_psd}(x, col = hcl.colors(200, "Blue-Red 3"), ...)
}
\arguments{
\item{x}{an object of class bdp_dw_tv_psd}

\item{col}{choice of color, default hcl.colors(200, "Blue-Red 3").}

\item{...}{further arguments to be parsed to \code{image.default}}
}
\description{
Plot method for bdp_dw_tv_psd class
}
\details{
Visualizes the spectral density function of time-varying
}
