% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Help function for visualizing}
\usage{
plotMCMC(mcmc, lambda, psd.true, ex.kurt.true = ex.kurt.true)
}
\description{
Help function for visualizing
}
\keyword{internal}
