% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{plotPsdEstimate}
\alias{plotPsdEstimate}
\title{Help function for visualizing}
\usage{
plotPsdEstimate(mcmc, lambda, psd.true)
}
\description{
Help function for visualizing
}
\keyword{internal}

