% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_S3method.R
\name{print.bdp_dw_result}
\alias{print.bdp_dw_result}
\title{Print method for bdp_dw_result class}
\usage{
\method{print}{bdp_dw_result}(x, ...)
}
\arguments{
\item{x}{object of class bdp_dw_result}

\item{...}{not in use}
}
\description{
Print method for bdp_dw_result class
}
