% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_S3.R
\name{print.gibbs_psd}
\alias{print.gibbs_psd}
\title{Print method for gibbs_psd class}
\usage{
\method{print}{gibbs_psd}(x, ...)
}
\arguments{
\item{x}{object of class gibbs_psd}

\item{...}{not in use}
}
\description{
Print method for gibbs_psd class
}
