% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_S3.R
\name{print_summary_gibbs_psd_help}
\alias{print_summary_gibbs_psd_help}
\title{Helping function for print and summary (both are quite similar)}
\usage{
print_summary_gibbs_psd_help(mcmc, flag = T)
}
\arguments{
\item{flag=T:}{print, flag=F: summary}
}
\description{
Helping function for print and summary (both are quite similar)
}
\keyword{internal}
