% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psd_arma}
\alias{psd_arma}
\title{Compute the ARMA(p,q) spectral density}
\usage{
psd_arma(freq, ar, ma, sigma2 = 1)
}
\arguments{
\item{freq}{numeric vector of frequencies to evaluate the psd, 0 <= freq < pi}

\item{ar}{autoregressive coefficients of ARMA model (use numeric(0) for empty AR part)}

\item{ma}{moving average coefficients of ARMA model (use numeric(0) for empty MA part)}

\item{sigma2}{the model innovation variance}
}
\value{
numeric vector of the (real-valued) spectral density values
}
\description{
Compute the ARMA(p,q) spectral density evaluated at some frequencies freq in [0,pi),
Note that no test for model stationarity is performed.
}

