% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_S3.R
\name{psd_array}
\alias{psd_array}
\title{Convert psd vector to array
(compatibility: to use plotMPsd for univariate functions as well)}
\usage{
psd_array(f)
}
\description{
Convert psd vector to array
(compatibility: to use plotMPsd for univariate functions as well)
}
\keyword{internal}
