% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{psd_varma_help}
\alias{psd_varma_help}
\title{helping function for psd_varma}
\usage{
psd_varma_help(freq, ar = matrix(nrow = nrow(sigma), ncol = 0),
  ma = matrix(nrow = nrow(sigma), ncol = 0), sigma)
}
\description{
helping function for psd_varma
}
\keyword{internal}
