% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{PSIwgt}
\alias{PSIwgt}
\title{Psi-weight calculation for a VARMA model.
NOTE: This is an exact copy of the MTS::PSIwgt function
(only with the plot functionality removed, as not needed).
This has to be done because the MTS package has been removed
from CRAN in April 2022.}
\usage{
PSIwgt(Phi = NULL, Theta = NULL, lag = 12, plot = TRUE, output = FALSE)
}
\description{
Psi-weight calculation for a VARMA model.
NOTE: This is an exact copy of the MTS::PSIwgt function
(only with the plot functionality removed, as not needed).
This has to be done because the MTS package has been removed
from CRAN in April 2022.
}
\keyword{internal}
