% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{qpsd}
\alias{qpsd}
\title{Compute a PSD in the Bernstein-Dirichlet parametrization.}
\usage{
qpsd(omega, v, w, k, beta_basis_k, epsilon = 1e-20)
}
\description{
Compute a PSD in the Bernstein-Dirichlet parametrization.
}
\keyword{internal}
