% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_miscellanea_zigzag.R
\name{qpsd_dw}
\alias{qpsd_dw}
\title{Evaluation of normalized time-varying spectral density function (based on posterior samples)}
\usage{
qpsd_dw(v, w1, w2, k1, k2, beta_basis_1_k, beta_basis_2_k)
}
\description{
Evaluation of normalized time-varying spectral density function (based on posterior samples)
}
\keyword{internal}
