% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_miscellanea_zigzag.R
\name{qpsd_dw.tilde_zigzag_cpp_expedited}
\alias{qpsd_dw.tilde_zigzag_cpp_expedited}
\title{Evaluation of normalized time-varying spectral density function (for MCMC algorithm)}
\usage{
qpsd_dw.tilde_zigzag_cpp_expedited(
  tilde.v,
  tilde.w1,
  tilde.w2,
  k1,
  k2,
  beta_basis_1_k,
  beta_basis_2_k
)
}
\description{
Evaluation of normalized time-varying spectral density function (for MCMC algorithm)
}
\keyword{internal}
