% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{realValuedPsd}
\alias{realValuedPsd}
\title{Store imaginary parts above and real parts below the diagonal}
\usage{
realValuedPsd(f_)
}
\description{
Store imaginary parts above and real parts below the diagonal
}
\keyword{internal}
