% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
rmvnorm(n, d, mu = rep(0, d), Sigma = diag(d), ...)
}
\arguments{
\item{n}{sample size}

\item{d}{dimensionality}

\item{mu}{mean vector}

\item{Sigma}{covariance matrix}

\item{...}{further arguments to be parsed to}
}
\value{
If n=1 a vector of length d, otherwise an n by d matrix with one sample in each row.
}
\description{
Produces one or more samples from the specified multivariate normal distribution.
}
\details{
This is a simple wrapper function based on \link[MASS]{mvrnorm},
to be used within \link[beyondWhittle]{sim_varma}
}
