% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sldmvnorm}
\alias{sldmvnorm}
\title{sum of multivariate normal log densities
with mean 0 and covariance Sigma, unnormalized}
\usage{
sldmvnorm(z_t, Sigma)
}
\description{
sum of multivariate normal log densities
with mean 0 and covariance Sigma, unnormalized
}
\keyword{internal}
