% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_S3method.R
\name{summary.bdp_dw_result}
\alias{summary.bdp_dw_result}
\title{Summary method for bdp_dw_result class}
\usage{
\method{summary}{bdp_dw_result}(object, ...)
}
\arguments{
\item{object}{object of class bdp_dw_result}

\item{...}{not in use}
}
\description{
Summary method for bdp_dw_result class
}
