% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_S3.R
\name{summary.gibbs_psd}
\alias{summary.gibbs_psd}
\title{Summary method for gibbs_psd class}
\usage{
\method{summary}{gibbs_psd}(object, ...)
}
\arguments{
\item{object}{object of class gibbs_psd}

\item{...}{not in use}
}
\description{
Summary method for gibbs_psd class
}
