% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{uci_matrix}
\alias{uci_matrix}
\title{Uniform credible intervals in matrix-valued case}
\usage{
uci_matrix(fpsd.sample, alpha, uniform_among_components = F)
}
\description{
Uniform credible intervals in matrix-valued case
}
\details{
see (6.5) in Meier (2018)
}
\keyword{internal}
