% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{uniformmax}
\alias{uniformmax}
\title{Uniform maximum, as needed for uniform credible intervals}
\usage{
uniformmax(sample)
}
\description{
Uniform maximum, as needed for uniform credible intervals
}
\details{
see Section 4.1 in Kirch et al (2018)
}
\keyword{internal}
