% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_I_l}
\alias{unit_trace_I_l}
\title{Range intervals I_l, see (63) in Mittelbach et al.}
\usage{
unit_trace_I_l(l)
}
\description{
Range intervals I_l, see (63) in Mittelbach et al.
}
\keyword{internal}
