% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unit_trace_L_from_x}
\alias{unit_trace_L_from_x}
\title{Get L (lower triangular Cholesky) from x
Called U^* in Mittelbach et al, see (60) there}
\usage{
unit_trace_L_from_x(x)
}
\description{
Get L (lower triangular Cholesky) from x
Called U^* in Mittelbach et al, see (60) there
}
\keyword{internal}
