% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_log_c}
\alias{unit_trace_log_c}
\title{Get log(c) vector, see (70) in Mittelbach et al.
Helping function for \code{unit_trace_runif}}
\usage{
unit_trace_log_c(p, q)
}
\description{
Get log(c) vector, see (70) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
}
\keyword{internal}
