% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_log_d}
\alias{unit_trace_log_d}
\title{Get log(d) vector, see (39) in Mittelbach et al, adjusted to complex case
Helping function for \code{unit_trace_runif}}
\usage{
unit_trace_log_d(p, q)
}
\description{
Get log(d) vector, see (39) in Mittelbach et al, adjusted to complex case
Helping function for \code{unit_trace_runif}
}
\keyword{internal}
