% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_log_f_l}
\alias{unit_trace_log_f_l}
\title{Get log(f_l), see (66) in Mittelbach et al.
Helping function for \code{unit_trace_runif}}
\usage{
unit_trace_log_f_l(phi, p, q, log_c, l)
}
\description{
Get log(f_l), see (66) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
}
\keyword{internal}
