% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_mu}
\alias{unit_trace_mu}
\title{Get mu vector, see (36) in Mittelbach et al.
Helping function for \code{unit_trace_runif}}
\usage{
unit_trace_mu(p, q)
}
\description{
Get mu vector, see (36) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
}
\keyword{internal}
