% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_nu}
\alias{unit_trace_nu}
\title{Get log(nu) vector, see (38) in Mittelbach et al.
Helping function for \code{unit_trace_runif}}
\usage{
unit_trace_nu(sigma2_vec, log_c_vec, log_d_vec)
}
\description{
Get log(nu) vector, see (38) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
}
\keyword{internal}
