% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_runif}
\alias{unit_trace_runif}
\title{Draw uniformly from Hpd matrices with unit trace}
\usage{
unit_trace_runif(n, d, verbose = F)
}
\description{
Draw uniformly from Hpd matrices with unit trace
}
\keyword{internal}
