% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_runif_single}
\alias{unit_trace_runif_single}
\title{Obtain one uniform draw from d times d Hpd matrices with unit trace
See Algorithm 2 in Mittelbach et al. (adjusted to complex case)}
\usage{
unit_trace_runif_single(d)
}
\description{
Obtain one uniform draw from d times d Hpd matrices with unit trace
See Algorithm 2 in Mittelbach et al. (adjusted to complex case)
}
\keyword{internal}
