% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_sigma2}
\alias{unit_trace_sigma2}
\title{Get sigma2 vector, see (70) in Mittelbach et al.
Helping function for \code{unit_trace_runif}}
\usage{
unit_trace_sigma2(p, q)
}
\description{
Get sigma2 vector, see (70) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
}
\keyword{internal}
