% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_trace.R
\name{unit_trace_U_from_phi}
\alias{unit_trace_U_from_phi}
\title{Get U (Hpd with unit trace) matrix from 
phi (hyperspherical coordinates) vector.}
\usage{
unit_trace_U_from_phi(phi)
}
\description{
Get U (Hpd with unit trace) matrix from 
phi (hyperspherical coordinates) vector.
}
\keyword{internal}
