% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unit_trace_x_from_phi}
\alias{unit_trace_x_from_phi}
\title{Get x from phi, see (62) in Mittelbach et al.}
\usage{
unit_trace_x_from_phi(phi)
}
\description{
Get x from phi, see (62) in Mittelbach et al.
}
\keyword{internal}
