% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unrollPsd}
\alias{unrollPsd}
\title{Redundantly roll out a PSD from length N=floor(n/2) to length n}
\usage{
unrollPsd(qPsd, n)
}
\description{
Redundantly roll out a PSD from length N=floor(n/2) to length n
}
\keyword{internal}
