% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{VAR_regressor_matrix}
\alias{VAR_regressor_matrix}
\title{VAR regressor matrix, see Section 2.2.3 in Koop and Korobilis (2010)}
\usage{
VAR_regressor_matrix(data, var.order)
}
\description{
VAR regressor matrix, see Section 2.2.3 in Koop and Korobilis (2010)
}
\keyword{internal}
