% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varma.R
\name{VARMAcov_muted}
\alias{VARMAcov_muted}
\title{This is a nearly exact copy of the MTS::VARMAcov function, where 
the output commands at the end are removed.
This has to be done because the function is called repeatedly
within the MCMC algorithm.
For future versions of the package, a better solution is intended.}
\usage{
VARMAcov_muted(Phi = NULL, Theta = NULL, Sigma = NULL, lag = 12,
  trun = 120)
}
\description{
This is a nearly exact copy of the MTS::VARMAcov function, where 
the output commands at the end are removed.
This has to be done because the function is called repeatedly
within the MCMC algorithm.
For future versions of the package, a better solution is intended.
}
\keyword{internal}
