% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vFromP}
\alias{vFromP}
\title{Get v from p (DP inverse stick breaking)
Note: p is assumed to have length L, i.e. it does NOT contain p_0}
\usage{
vFromP(p, eps = 1e-08)
}
\description{
Get v from p (DP inverse stick breaking)
Note: p is assumed to have length L, i.e. it does NOT contain p_0
}
\keyword{internal}
