% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acvMatrix}
\alias{acvMatrix}
\title{C++ function to build a Toeplitz ACV matric, given an ACV vector.}
\usage{
acvMatrix(acv)
}
\description{
C++ function to build a Toeplitz ACV matric, given an ACV vector.
}
\keyword{internal}

