% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{arma_conditional}
\alias{arma_conditional}
\title{Function for computing the ARMA(p, q) conditional likelihood
Models with AR component: zt are from t = p + 1 to m
Models with MA component: eps_0 = ... = eps_{q+1} = 0
The input zt should be FCFZ, i.e., data corrected in freq. domain then IFTd
FCFZ should be FCFZ[-c(1, n)]; i.e., have removed elements 1 and n beforehand
No sigma2 in here anymore!
...: Further arguments to be passed to likelihood function}
\usage{
arma_conditional(zt, ar, ma, nll_fun, sigma2, ...)
}
\description{
Function for computing the ARMA(p, q) conditional likelihood
Models with AR component: zt are from t = p + 1 to m
Models with MA component: eps_0 = ... = eps_{q+1} = 0
The input zt should be FCFZ, i.e., data corrected in freq. domain then IFTd
FCFZ should be FCFZ[-c(1, n)]; i.e., have removed elements 1 and n beforehand
No sigma2 in here anymore!
...: Further arguments to be passed to likelihood function
}
\keyword{internal}

