% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{coarsened_bernstein}
\alias{coarsened_bernstein}
\title{Help function for polynomial basis.}
\usage{
coarsened_bernstein(omega, l)
}
\description{
Help function for polynomial basis.
}
\keyword{internal}
