% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{coarsened_bernstein_i}
\alias{coarsened_bernstein_i}
\title{Help function for polynomial basis.}
\usage{
coarsened_bernstein_i(omega, l, i)
}
\description{
Help function for polynomial basis.
}
\keyword{internal}
