% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{dbList}
\alias{dbList}
\title{Help function for polynomial basis.}
\usage{
dbList(n, kmax, bernstein_l = 0, bernstein_r = 1, coarsened = F)
}
\description{
Help function for polynomial basis.
}
\keyword{internal}
