% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_util.R
\name{fast_ift}
\alias{fast_ift}
\title{Compute F_n^t X_n with the real-valued Fourier matrix F_n}
\usage{
fast_ift(x)
}
\description{
Compute F_n^t X_n with the real-valued Fourier matrix F_n
}
\keyword{internal}
