% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genEpsARMAC}
\alias{genEpsARMAC}
\title{C++ function for generating epsilon process for ARMA(p,q)}
\usage{
genEpsARMAC(zt, ar, ma)
}
\description{
C++ function for generating epsilon process for ARMA(p,q)
}
\keyword{internal}

