% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{llike}
\alias{llike}
\title{Log corrected parametric likelihood}
\usage{
llike(omega, FZ, ar, ma, v, w, k, tau, corrected, toggle.q, pdgrm, beta_basis_k,
  nll_fun, f.alpha, excludeBoundary, full_lik)
}
\description{
Log corrected parametric likelihood
}
\keyword{internal}
