% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{llike}
\alias{llike}
\title{Log corrected parametric likelihood}
\usage{
llike(omega, FZ, ar, ma, v, w, k, tau, corrected, toggle.q, pdgrm, db.list,
  dist, nll_fun, ex.kurt, beta, f.alpha)
}
\description{
Log corrected parametric likelihood
}
\keyword{internal}
