% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_np_help.R
\name{logDet_stickBreaking}
\alias{logDet_stickBreaking}
\title{Log determinant of stick breaking transformation V -> p}
\usage{
logDet_stickBreaking(v)
}
\description{
Log determinant of stick breaking transformation V -> p
}
\details{
See Section 3 in Choudhuri et al (2004)
}
\keyword{internal}
