% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern__gibbs_core.R
\name{lpost}
\alias{lpost}
\title{Log corrected posterior}
\usage{
lpost(omega, FZ, ar, ma, v, w, k, tau, M, g0.alpha, g0.beta, k.theta, tau.alpha,
  tau.beta, corrected, toggle.q, pdgrm, db.list, dist, nll_fun, ex.kurt,
  kurt.lambda, beta, beta.alpha, beta.beta, f.alpha, rho, rho.alpha, rho.beta)
}
\description{
Log corrected posterior
}
\keyword{internal}
