% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_ar_help.R
\name{lpost_AR}
\alias{lpost_AR}
\title{Log Posterior = Log Prior + (conditional) Log Likelihood}
\usage{
lpost_AR(x_noise, rho, rho.alpha, rho.beta, sigma2, sigma2.alpha,
  sigma2.beta, full_lik)
}
\description{
Log Posterior = Log Prior + (conditional) Log Likelihood
}
\keyword{internal}
