% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_np_help.R
\name{lprior}
\alias{lprior}
\title{Log prior of Bernstein-Dirichlet mixture and parametric working model -- all unnormalized}
\usage{
lprior(v, w, k, tau, M, g0.alpha, g0.beta, k.theta, tau.alpha, tau.beta,
  f.alpha, corrected, rho, rho.alpha, rho.beta)
}
\description{
Log prior of Bernstein-Dirichlet mixture and parametric working model -- all unnormalized
}
\details{
See Section 3 in Kirch et al (2018).
Hyperparameters are M, g0.a, g0.b, k.theta, tau.alpha, tau.beta.
Note: Flat prior on f.alpha.
}
\keyword{internal}
