% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_ar_help.R
\name{lprior_AR}
\alias{lprior_AR}
\title{Log prior for PACF (~Beta) and sigma2 (~InverseGamma), unnormalized}
\usage{
lprior_AR(rho, rho.alpha, rho.beta, sigma2, sigma2.alpha, sigma2.beta)
}
\description{
Log prior for PACF (~Beta) and sigma2 (~InverseGamma), unnormalized
}
\keyword{internal}
