% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicWhittle_miscellanea_zigzag.R
\name{lprior_dw}
\alias{lprior_dw}
\title{Calculation of log prior}
\usage{
lprior_dw(
  tilde.v,
  tilde.w1,
  tilde.w2,
  k1,
  k2,
  tau,
  M,
  g0.alpha,
  g0.beta,
  k1.theta,
  k2.theta,
  tau.alpha,
  tau.beta
)
}
\description{
Calculation of log prior
}
\keyword{internal}
