% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_np_help.R
\name{my_rdirichlet}
\alias{my_rdirichlet}
\title{Generate a random samples from a Dirichlet distribution}
\usage{
my_rdirichlet(alpha)
}
\arguments{
\item{alpha}{numeric vector of positive concentration parameter}
}
\value{
a vector of the same length as alpha
}
\description{
Generate a random samples from a Dirichlet distribution
}
\keyword{internal}
