% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_np_help.R
\name{qpsd}
\alias{qpsd}
\title{Compute normalized PSD in the Bernstein-Dirichlet parametrization.}
\usage{
qpsd(omega, v, w, k, beta_basis_k, epsilon = 1e-20)
}
\description{
Compute normalized PSD in the Bernstein-Dirichlet parametrization.
}
\details{
See (5) in Choudhuri et al (2004)
}
\keyword{internal}
