% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vFromP}
\alias{vFromP}
\title{C++ function for generating  v from p (inverse stick breaking)
NOTE: p is assumed to have length L, i.e. it does NOT contain p_0 !!}
\usage{
vFromP(p, eps = 1e-08)
}
\description{
C++ function for generating  v from p (inverse stick breaking)
NOTE: p is assumed to have length L, i.e. it does NOT contain p_0 !!
}
\keyword{internal}
